part 4 done

+++ part4/blog-list/controllers/blogs.js
@@ -1,7 +1,9 @@
 const blogsRouter = require('express').Router()
 const Blog = require('../models/blog')
 const User = require('../models/user')
-
+const jwt = require('jsonwebtoken')
+const config = require('../utils/config')
+require('express-async-errors')
 
 blogsRouter.get('/', async (request, response) => {
 	const blogs = await Blog.find({}).populate('user', { username: 1, name: 1 })
@@ -10,13 +12,15 @@ blogsRouter.get('/', async (request, response) => {
 })
 
 blogsRouter.post('/', async (request, response) => {
-	// Get first user as demonstration
-	const user = await User.findOne({})
+	const decodedToken = jwt.verify(request.token, config.SECRET)
+	if (!request.token || !decodedToken.id || !decodedToken.username) {
+		return response.status(401).json({ error: 'token missing or invalid' })
+	}
+	const user = await User.findById(decodedToken.id)
 
-	const blog = new Blog({ user: user.id, ...request.body })
+	const blog = new Blog({ user: user._id, ...request.body })
 
 	const savedBlog = await blog.save()
-
 	user.blogs = user.blogs.concat(savedBlog._id)
 	await user.save()
 
@@ -24,6 +28,16 @@ blogsRouter.post('/', async (request, response) => {
 })
 
 blogsRouter.delete('/:id', async (request, response) => {
+	const decodedToken = jwt.verify(request.token, config.SECRET)
+	if (!request.token || !decodedToken.id || !decodedToken.username) {
+		return response.status(401).json({ error: 'token missing or invalid' })
+	}
+
+	const blog = await Blog.findById(request.params.id)
+	if (blog.user.toString() !== decodedToken.id.toString()) {
+		return response.status(401).json({ error: 'invalid permissions' })
+	}
+
 	await Blog.findByIdAndDelete(request.params.id)
 	response.status(204).end()
 })
